stop short-circuiting the SSP code for sspstrong
By returning early for sspstrong, this function is skipping the process
of assigning layouts for any of the alloca calls in the function. This
means local variables aren't being properly protected when this early
return path is triggered. With Clang, this appears to happen when either
variable length arrays or alloca are used, but not regular arrays.